+++ packages/lib-sourcify/src/SourcifyChain/SourcifyChainTypes.ts
@@ -41,6 +41,9 @@ export interface FetchContractCreationTxMethods {
     runtime: string;
   };
   veChainApi?: boolean;
+  nodeRealApi?: {
+    url: string;
+  };
 }
 
 // types of the keys of FetchContractCreationTxMethods