+++ services/server/test/unit/VerificationService.spec.ts
@@ -107,6 +107,7 @@ describe("VerificationService", function () {
         solcRepoPath: config.get("solcRepo"),
         solJsonRepoPath: config.get("solJsonRepo"),
         vyperRepoPath: config.get("vyperRepo"),
+        feRepoPath: config.get("feRepo"),
       },
       new StorageService({
         enabledServices: {
@@ -146,6 +147,7 @@ describe("VerificationService", function () {
         solcRepoPath: config.get("solcRepo"),
         solJsonRepoPath: config.get("solJsonRepo"),
         vyperRepoPath: config.get("vyperRepo"),
+        feRepoPath: config.get("feRepo"),
       },
       mockStorageService,
     );
@@ -206,6 +208,7 @@ describe("VerificationService", function () {
         solcRepoPath: config.get("solcRepo"),
         solJsonRepoPath: config.get("solJsonRepo"),
         vyperRepoPath: config.get("vyperRepo"),
+        feRepoPath: config.get("feRepo"),
         debugDataS3Config: {
           bucket: testS3Bucket,
           region: "test-region",
@@ -261,6 +264,7 @@ describe("VerificationService", function () {
         solcRepoPath: config.get("solcRepo"),
         solJsonRepoPath: config.get("solJsonRepo"),
         vyperRepoPath: config.get("vyperRepo"),
+        feRepoPath: config.get("feRepo"),
         debugDataS3Config: {
           bucket: testS3Bucket,
           region: "test-region",