+++ packages/lib-sourcify/CHANGELOG.md
@@ -2,11 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## @ethereum-sourcify/lib-sourcify@3.4.0 - 2026-03-30
+
+- feat: add Fe language support for smart contract verification (#2692)
+- feat: add NodeReal nr_getContractCreationTransaction for BSC mainnet (#2702)
+
 ## @ethereum-sourcify/lib-sourcify@3.3.0 - 2026-03-17
 
 - Support all Solidity versions >= 0.1.3 (#2652)
 - Add transientStorageLayout to compiler output (#2689)
-
 
 ## @ethereum-sourcify/lib-sourcify@3.2.0 - 2026-03-02
 
@@ -22,7 +26,6 @@ All notable changes to this project will be documented in this file.
 ## @ethereum-sourcify/lib-sourcify@3.1.0 - 2026-01-07
 
 - Add support for Yul compilation and verification (#2521)
-
 
 ## @ethereum-sourcify/lib-sourcify@3.0.5 - 2025-12-15
 
@@ -42,7 +45,6 @@ All notable changes to this project will be documented in this file.
 ## @ethereum-sourcify/lib-sourcify@3.0.2 - 2025-10-30
 
 - Only count timeouts and network errors as rpc failures (#2473)
-
 
 ## @ethereum-sourcify/lib-sourcify@3.0.1 - 2025-10-28